move load more button to the bottom of the nugget list
